Split data fetching and presentation
